+++ L2J_Mobius_CT_0_Interlude/java/org/l2jmobius/gameserver/model/holders/EventDropHolder.java
@@ -27,7 +27,7 @@ public class EventDropHolder extends DropHolder
 	private final int _maxLevel;
 	private final Set<Integer> _monsterIds;
 	
-	public EventDropHolder(int itemId, long min, long max, double chance, int minLevel, int maxLevel, Set<Integer> monsterIds)
+	public EventDropHolder(int itemId, int min, int max, double chance, int minLevel, int maxLevel, Set<Integer> monsterIds)
 	{
 		super(null, itemId, min, max, chance);
 		_minLevel = minLevel;
